Move Product lightbox logic to reusable HOC.
