+++ client/a8c-for-agencies/sections/marketplace/products-overview-v2/product-card/index.tsx
@@ -3,38 +3,26 @@ import { Button } from '@wordpress/components';
 import { check } from '@wordpress/icons';
 import clsx from 'clsx';
 import { useTranslate } from 'i18n-calypso';
-import { useCallback, useEffect, useMemo, useState } from 'react';
-import { useDispatch } from 'react-redux';
-import {
-	useProductDescription,
-	useURLQueryParams,
-} from 'calypso/jetpack-cloud/sections/partner-portal/hooks';
-import { LICENSE_INFO_MODAL_ID } from 'calypso/jetpack-cloud/sections/partner-portal/lib';
+import { useCallback, useEffect, useMemo } from 'react';
+import { useProductDescription } from 'calypso/jetpack-cloud/sections/partner-portal/hooks';
 import getProductShortTitle from 'calypso/jetpack-cloud/sections/partner-portal/lib/get-product-short-title';
-import LicenseLightbox from 'calypso/jetpack-cloud/sections/partner-portal/license-lightbox';
 import LicenseLightboxLink from 'calypso/jetpack-cloud/sections/partner-portal/license-lightbox-link';
-import { recordTracksEvent } from 'calypso/state/analytics/actions';
-import { APIProductFamilyProduct } from '../../../../../state/partner-portal/types';
+import withProductLightbox, {
+	ProductLightboxActivatorProps,
+	WithProductLightboxProps,
+} from '../hocs/with-product-lightbox';
 import ProductBadges from '../product-badges';
 import ProductPriceWithDiscount from './product-price-with-discount-info';
-import WooPaymentsCustomDescription from './woopayments-custom-description';
-import WooPaymentsCustomFooter from './woopayments-custom-footer';
-import WooPaymentsRevenueShareNotice from './woopayments-revenue-share-notice';
 
 import './style.scss';
 
-interface Props {
-	asReferral?: boolean;
-	product: APIProductFamilyProduct;
-	isSelected: boolean;
-	isDisabled?: boolean;
-	onSelectProduct: ( value: APIProductFamilyProduct ) => void | null;
-	suggestedProduct?: string | null;
-	hideDiscount?: boolean;
-	quantity?: number;
-}
+type Props = WithProductLightboxProps &
+	ProductLightboxActivatorProps & {
+		suggestedProduct?: string | null;
+		hideDiscount?: boolean;
+	};
 
-export default function ProductCard( props: Props ) {
+function ProductCard( props: Props ) {
 	const {
 		asReferral,
 		product,
@@ -44,15 +32,11 @@ export default function ProductCard( props: Props ) {
 		suggestedProduct,
 		hideDiscount,
 		quantity,
+		onShowLightbox,
 	} = props;
-
-	const { setParams, resetParams, getParamValue } = useURLQueryParams();
-	const modalParamValue = getParamValue( LICENSE_INFO_MODAL_ID );
 	const productTitle = getProductShortTitle( product );
 
-	const [ showLightbox, setShowLightbox ] = useState( modalParamValue === product.slug );
 	const translate = useTranslate();
-	const dispatch = useDispatch();
 
 	const onSelect = useCallback( () => {
 		if ( isDisabled ) {
@@ -60,7 +44,7 @@ export default function ProductCard( props: Props ) {
 		}
 
 		onSelectProduct?.( product );
-	}, [ onSelectProduct, product ] );
+	}, [ isDisabled, onSelectProduct, product ] );
 
 	const onKeyDown = useCallback(
 		( e: any ) => {
@@ -95,32 +79,6 @@ export default function ProductCard( props: Props ) {
 
 	const { description: productDescription } = useProductDescription( product.slug );
 
-	const onShowLightbox = useCallback(
-		( e: React.MouseEvent< HTMLElement > ) => {
-			e.stopPropagation();
-
-			dispatch(
-				recordTracksEvent( 'calypso_marketplace_products_overview_product_view', {
-					product: product.slug,
-				} )
-			);
-
-			setParams( [
-				{
-					key: LICENSE_INFO_MODAL_ID,
-					value: product.slug,
-				},
-			] );
-			setShowLightbox( true );
-		},
-		[ dispatch, product.slug, setParams ]
-	);
-
-	const onHideLightbox = useCallback( () => {
-		resetParams( [ LICENSE_INFO_MODAL_ID ] );
-		setShowLightbox( false );
-	}, [ resetParams ] );
-
 	const ctaLabel = useMemo( () => {
 		const selectedQuantity = quantity ?? 1;
 
@@ -137,121 +95,63 @@ export default function ProductCard( props: Props ) {
 		return isSelected ? translate( 'Added to cart' ) : translate( 'Add to cart' );
 	}, [ asReferral, isSelected, quantity, translate ] );
 
-	const customDescription = useMemo( () => {
-		if ( product.slug === 'woocommerce-woopayments' ) {
-			return <WooPaymentsCustomDescription />;
-		}
-
-		return undefined;
-	}, [ product.slug ] );
-
-	const customFooter = useMemo( () => {
-		if ( product.slug === 'woocommerce-woopayments' ) {
-			return <WooPaymentsCustomFooter />;
-		}
-
-		return undefined;
-	}, [ product.slug ] );
-
-	const extraAsideContent = useMemo( () => {
-		if ( product.slug === 'woocommerce-woopayments' ) {
-			return <WooPaymentsRevenueShareNotice />;
-		}
-
-		return undefined;
-	}, [ product.slug ] );
-
-	const ctaLightboxLabel = useMemo( () => {
-		const selectedQuantity = quantity ?? 1;
-
-		if ( asReferral ) {
-			return isSelected ? translate( 'Remove from referral' ) : translate( 'Add to referral' );
-		}
-
-		if ( selectedQuantity > 1 ) {
-			return isSelected
-				? translate( 'Remove %(quantity)s from cart', { args: { quantity: selectedQuantity } } )
-				: translate( 'Add %(quantity)s to cart', { args: { quantity: selectedQuantity } } );
-		}
-
-		return isSelected ? translate( 'Remove from cart' ) : translate( 'Add to cart' );
-	}, [ asReferral, isSelected, quantity, translate ] );
-
 	const isRedesign = isEnabled( 'a4a-product-page-redesign' );
 
 	return (
-		<>
-			<div
-				onClick={ onSelect }
-				onKeyDown={ onKeyDown }
-				role="button"
-				tabIndex={ 0 }
-				aria-disabled={ isDisabled }
-				className={ clsx( {
-					'product-card': true,
-					selected: isSelected,
-					disabled: isDisabled,
-				} ) }
-			>
-				<div className="product-card__inner">
-					<div className="product-card__details">
-						<div className="product-card__main">
-							<div className="product-card__heading">
-								<h3 className="product-card__title">{ productTitle }</h3>
-								<ProductBadges product={ product } />
-								<div className="product-card__pricing is-compact">
-									<ProductPriceWithDiscount
-										product={ product }
-										hideDiscount={ hideDiscount }
-										quantity={ quantity }
-										compact
-									/>
-								</div>
+		<div
+			onClick={ onSelect }
+			onKeyDown={ onKeyDown }
+			role="button"
+			tabIndex={ 0 }
+			aria-disabled={ isDisabled }
+			className={ clsx( {
+				'product-card': true,
+				selected: isSelected,
+				disabled: isDisabled,
+			} ) }
+		>
+			<div className="product-card__inner">
+				<div className="product-card__details">
+					<div className="product-card__main">
+						<div className="product-card__heading">
+							<h3 className="product-card__title">{ productTitle }</h3>
+							<ProductBadges product={ product } />
+							<div className="product-card__pricing is-compact">
+								<ProductPriceWithDiscount
+									product={ product }
+									hideDiscount={ hideDiscount }
+									quantity={ quantity }
+									compact
+								/>
+							</div>
 
-								<div className="product-card__description">
-									{ truncateDescription( productDescription ) }
-								</div>
+							<div className="product-card__description">
+								{ truncateDescription( productDescription ) }
 							</div>
 						</div>
 					</div>
-					<div className="product-card__buttons">
-						<Button
-							className={ clsx( { 'product-card__select-button': ! isRedesign } ) }
-							variant={ ! isSelected ? 'primary' : 'secondary' }
-							tabIndex={ -1 }
-							icon={ isSelected ? check : undefined }
-						>
-							{ ctaLabel }
-						</Button>
-						{ ! /^jetpack-backup-addon-storage-/.test( product.slug ) && (
-							<LicenseLightboxLink
-								customText={ translate( 'View details' ) }
-								productName={ productTitle }
-								onClick={ onShowLightbox }
-								showIcon={ ! isRedesign }
-							/>
-						) }
-					</div>
+				</div>
+				<div className="product-card__buttons">
+					<Button
+						className={ clsx( { 'product-card__select-button': ! isRedesign } ) }
+						variant={ ! isSelected ? 'primary' : 'secondary' }
+						tabIndex={ -1 }
+						icon={ isSelected ? check : undefined }
+					>
+						{ ctaLabel }
+					</Button>
+					{ ! /^jetpack-backup-addon-storage-/.test( product.slug ) && (
+						<LicenseLightboxLink
+							customText={ translate( 'View details' ) }
+							productName={ productTitle }
+							onClick={ onShowLightbox }
+							showIcon={ ! isRedesign }
+						/>
+					) }
 				</div>
 			</div>
-			{ showLightbox && (
-				<LicenseLightbox
-					product={ product }
-					quantity={ quantity }
-					ctaLabel={ ctaLightboxLabel as string }
-					isCTAPrimary={ ! isSelected }
-					isDisabled={ isDisabled }
-					onActivate={ onSelectProduct }
-					onClose={ onHideLightbox }
-					customDescription={ customDescription }
-					customFooter={ customFooter }
-					extraAsideContent={ extraAsideContent }
-				/>
-			) }
-		</>
+		</div>
 	);
 }
 
-ProductCard.defaultProps = {
-	onSelectProduct: null,
-};
+export default withProductLightbox( ProductCard );
